+++ simulate/simulate.h
@@ -47,7 +47,7 @@ class Simulate {
   using Clock = std::chrono::steady_clock;
   static_assert(std::ratio_less_equal_v<Clock::period, std::milli>);
 
-  static constexpr int kMaxGeom = 20000;
+  static constexpr int kMaxGeom = 100000;
 
   // create object and initialize the simulate ui
   Simulate(