+++ doc/changelog.rst
@@ -14,11 +14,16 @@ General
   as input to learning-based agents and in environment logic.
 - Removed the SdfLib plugin and the dependency on `SdfLib <https://github.com/UPC-ViRVIG/SdfLib>`__. SDFs are now
   supported natively in mjModel.
+- Removed ``oct_depth`` from :ref:`mjvOption` (unused).
 - Inertia computation in MuJoCo C is now performed by a new :ref:`pipeline<piStages>` function :ref:`mj_makeM`, which
   combines the Composite Rigid Body algorithm in :ref:`mj_crb` and additional terms related to
   :ref:`tendon armature<tendon-spatial-armature>`. Code that uses :ref:`mj_crb` to compute the inertia should now use
   :ref:`mj_makeM` instead.
 
+.. admonition:: Breaking API changes
+   :class: attention
+
+   - Removed the ``mjVIS_FLEXBVH`` enum value, its functionality is now provided by :ref:`mjVIS_MESHBVH<mjtVisFlag>`.
 
 Bug fixes
 ^^^^^^^^^